Moved Elasticsearch client to registry-common project
