+++ pom.xml
@@ -45,11 +45,11 @@
             <artifactId>commons-cli</artifactId>
             <version>1.4</version>
         </dependency>
-        <!-- Elasticsearch client-->
+        <!-- Registry Common library-->
         <dependency>
-            <groupId>org.elasticsearch.client</groupId>
-            <artifactId>elasticsearch-rest-client</artifactId>
-            <version>7.8.1</version>
+            <groupId>gov.nasa.pds</groupId>
+            <artifactId>registry-common</artifactId>
+            <version>1.0.0</version>
         </dependency>
         <!-- JSON parser -->
         <dependency>
