+++ src/main/java/gov/nasa/pds/registry/mgr/cmd/data/SetArchiveStatusCmd.java
@@ -15,12 +15,12 @@ import org.elasticsearch.client.RestClient;
 
 import com.google.gson.Gson;
 
+import gov.nasa.pds.registry.common.es.client.EsClientFactory;
+import gov.nasa.pds.registry.common.es.client.EsUtils;
 import gov.nasa.pds.registry.mgr.Constants;
 import gov.nasa.pds.registry.mgr.cmd.CliCommand;
-import gov.nasa.pds.registry.mgr.es.client.EsClientFactory;
 import gov.nasa.pds.registry.mgr.util.CloseUtils;
 import gov.nasa.pds.registry.mgr.util.es.EsRequestBuilder;
-import gov.nasa.pds.registry.mgr.util.es.EsUtils;
 
 
 public class SetArchiveStatusCmd implements CliCommand
